requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
defusedxml==0.7.1
Django==4.2.11
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.1
httpie==3.2.2
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
psycopg==3.1.18
Pygments==2.18.0
pymemcache==4.0.0
PySocks==1.7.1
redis==5.0.3
requests==2.32.2
requests-toolbelt==1.0.0
rich==13.7.1
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1